Fixed include dependencies Removed ramdisk.c, defs.h
