+++ kernel/bio.c
@@ -17,11 +17,9 @@
 #include "types.h"
 #include "param.h"
 #include "spinlock.h"
-#include "sleeplock.h"
-#include "riscv.h"
-#include "defs.h"
-#include "fs.h"
 #include "buf.h"
+#include "virtio.h"
+#include "printf.h"
 
 struct {
   struct spinlock lock;
