+++ kernel/console.c
@@ -12,15 +12,11 @@
 #include <stdarg.h>
 
 #include "types.h"
-#include "param.h"
 #include "spinlock.h"
-#include "sleeplock.h"
-#include "fs.h"
 #include "file.h"
-#include "memlayout.h"
-#include "riscv.h"
-#include "defs.h"
+#include "console.h"
 #include "proc.h"
+#include "uart.h"
 
 #define BACKSPACE 0x100
 #define C(x)  ((x)-'@')  // Control-x
