+++ kernel/console.h
@@ -1,0 +1,5 @@
+#pragma once
+
+void consoleinit(void);
+void consoleintr(int);
+void consputc(int);
