+++ kernel/fs.h
@@ -1,5 +1,8 @@
 #pragma once
 
+#include "file.h"
+#include "stat.h"
+
 // On-disk file system format.
 // Both the kernel and user programs use this header file.
 
@@ -26,6 +29,7 @@ struct superblock {
 
 #define FSMAGIC 0x10203040
 
+// Remember to change this constant in file.h too
 #define NDIRECT 12
 #define NINDIRECT (BSIZE / sizeof(uint))
 #define MAXFILE (NDIRECT + NINDIRECT)
@@ -59,4 +63,24 @@ struct dirent {
   ushort inum;
   char name[DIRSIZ];
 };
+
+
+void            fsinit(int);
+int             dirlink(struct inode*, char*, uint);
+struct inode*   dirlookup(struct inode*, char*, uint*);
+struct inode*   ialloc(uint, short);
+struct inode*   idup(struct inode*);
+void            iinit();
+void            ilock(struct inode*);
+void            iput(struct inode*);
+void            iunlock(struct inode*);
+void            iunlockput(struct inode*);
+void            iupdate(struct inode*);
+int             namecmp(const char*, const char*);
+struct inode*   namei(char*);
+struct inode*   nameiparent(char*, char*);
+int             readi(struct inode*, int, uint64, uint, uint);
+void            stati(struct inode*, struct stat*);
+int             writei(struct inode*, int, uint64, uint, uint);
+void            itrunc(struct inode*);
