+++ kernel/kalloc.h
@@ -1,0 +1,5 @@
+#pragma once
+
+void* kalloc(void);
+void kfree(void *);
+void kinit(void);
