+++ kernel/log.c
@@ -1,11 +1,11 @@
-#include "types.h"
-#include "riscv.h"
-#include "defs.h"
 #include "param.h"
 #include "spinlock.h"
-#include "sleeplock.h"
 #include "fs.h"
 #include "buf.h"
+#include "log.h"
+#include "proc.h"
+#include "printf.h"
+#include "string.h"
 
 // Simple logging that allows concurrent FS system calls.
 //
