+++ kernel/log.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include "buf.h"
+
+void            initlog(int, struct superblock*);
+void            log_write(struct buf*);
+void            begin_op(void);
+void            end_op(void);
