+++ kernel/main.c
@@ -1,8 +1,11 @@
-#include "types.h"
-#include "param.h"
-#include "memlayout.h"
-#include "riscv.h"
-#include "defs.h"
+#include "console.h"
+#include "proc.h"
+#include "vm.h"
+#include "plic.h"
+#include "trap.h"
+#include "virtio.h"
+#include "kalloc.h"
+#include "printf.h"
 
 volatile static int started = 0;
