+++ kernel/pipe.c
@@ -1,23 +1,9 @@
 #include "types.h"
-#include "riscv.h"
-#include "defs.h"
-#include "param.h"
-#include "spinlock.h"
 #include "proc.h"
-#include "fs.h"
-#include "sleeplock.h"
 #include "file.h"
-
-#define PIPESIZE 512
-
-struct pipe {
-  struct spinlock lock;
-  char data[PIPESIZE];
-  uint nread;     // number of bytes read
-  uint nwrite;    // number of bytes written
-  int readopen;   // read fd is still open
-  int writeopen;  // write fd is still open
-};
+#include "pipe.h"
+#include "vm.h"
+#include "kalloc.h"
 
 int
 pipealloc(struct file **f0, struct file **f1)
