+++ kernel/plic.h
@@ -1,0 +1,6 @@
+#pragma once
+
+void            plicinit(void);
+void            plicinithart(void);
+int             plic_claim(void);
+void            plic_complete(int);
