+++ kernel/printf.h
@@ -1,0 +1,5 @@
+#pragma once
+
+void            printf(char*, ...);
+void            panic(char*) __attribute__((noreturn));
+void            printfinit(void);
