+++ kernel/sleeplock.c
@@ -1,10 +1,5 @@
 // Sleeping locks
 
-#include "types.h"
-#include "riscv.h"
-#include "defs.h"
-#include "param.h"
-#include "memlayout.h"
 #include "spinlock.h"
 #include "proc.h"
 #include "sleeplock.h"
