+++ kernel/sleeplock.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include "spinlock.h"
+
 // Long-term locks for processes
 struct sleeplock {
   uint locked;       // Is the lock held?
@@ -10,3 +12,7 @@ struct sleeplock {
   int pid;           // Process holding lock
 };
 
+void acquiresleep(struct sleeplock*);
+void releasesleep(struct sleeplock*);
+int holdingsleep(struct sleeplock*);
+void initsleeplock(struct sleeplock*, char*);
