+++ kernel/spinlock.c
@@ -1,12 +1,9 @@
 // Mutual exclusion spin locks.
 
-#include "types.h"
-#include "param.h"
-#include "memlayout.h"
 #include "spinlock.h"
 #include "riscv.h"
 #include "proc.h"
-#include "defs.h"
+#include "printf.h"
 
 void
 initlock(struct spinlock *lk, char *name)
