+++ kernel/spinlock.h
@@ -1,5 +1,9 @@
 #pragma once
 
+#include "types.h"
+
+struct cpu;
+
 // Mutual exclusion lock.
 struct spinlock {
   uint locked;       // Is the lock held?
@@ -9,3 +13,9 @@ struct spinlock {
   struct cpu *cpu;   // The cpu holding the lock.
 };
 
+void            acquire(struct spinlock*);
+int             holding(struct spinlock*);
+void            initlock(struct spinlock*, char*);
+void            release(struct spinlock*);
+void            push_off(void);
+void            pop_off(void);
