+++ kernel/start.c
@@ -2,7 +2,6 @@
 #include "param.h"
 #include "memlayout.h"
 #include "riscv.h"
-#include "defs.h"
 
 void main();
 void timerinit();
