+++ kernel/stat.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include "types.h"
+
 #define T_DIR     1   // Directory
 #define T_FILE    2   // File
 #define T_DEVICE  3   // Device
