+++ kernel/string.h
@@ -1,0 +1,13 @@
+#pragma once
+
+#include "types.h"
+
+#define NELEM(x) (sizeof(x)/sizeof((x)[0]))
+
+int             memcmp(const void*, const void*, uint);
+void*           memmove(void*, const void*, uint);
+void*           memset(void*, int, uint);
+char*           safestrcpy(char*, const char*, int);
+int             strlen(const char*);
+int             strncmp(const char*, const char*, uint);
+char*           strncpy(char*, const char*, int);
