+++ kernel/sysproc.c
@@ -1,10 +1,8 @@
 #include "types.h"
-#include "riscv.h"
-#include "defs.h"
-#include "param.h"
-#include "memlayout.h"
 #include "spinlock.h"
 #include "proc.h"
+#include "trap.h"
+#include "syscall.h"
 
 uint64
 sys_exit(void)
