+++ kernel/uart.h
@@ -1,0 +1,7 @@
+#pragma once
+
+void            uartinit(void);
+void            uartintr(void);
+void            uartputc(int);
+void            uartputc_sync(int);
+int             uartgetc(void);
