+++ kernel/virtio.h
@@ -1,5 +1,8 @@
 #pragma once
 
+#include "types.h"
+#include "buf.h"
+
 //
 // virtio device definitions.
 // for both the mmio interface, and virtio descriptors.
@@ -96,3 +99,8 @@ struct virtio_blk_req {
   uint32 reserved;
   uint64 sector;
 };
+
+void            virtio_disk_init(void);
+void            virtio_disk_rw(struct buf *, int);
+void            virtio_disk_intr(void);
+
