+++ user/initcode.S
@@ -1,7 +1,7 @@
 # Initial process that execs /init.
 # This code runs in user space.
 
-#include "syscall.h"
+#include "syscall_numbers.h"
 
 # exec(init, argv)
 .globl start
