+++ user/usys.pl
@@ -4,7 +4,7 @@
 
 print "# generated by usys.pl - do not edit\n";
 
-print "#include \"kernel/syscall.h\"\n";
+print "#include \"kernel/syscall_numbers.h\"\n";
 
 sub entry {
     my $name = shift;
